### 2025-05-30 #8318 @editorialbot
[PRE REVIEW]: stylo: a lightweight nf-core style nanopore assembly pipeline optimized for enteric bacteria
**Submitting author:** @mohitmthakur (<a href="http://orcid.org/0009-0003-2927-6051">Mohit Thakur</a>)
**Repository:** https://github.com/ncezid-biome/stylo
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
