### 2025-08-04 #8318 @arfon
@mohitmthakur – thanks for your submission to JOSS. Could you please point us to the software contributions of this project please?
