### 2025-08-21 9367f2a
changed medaka model selection, added model logging to outputs
