### 2025-12-26 #9695 @crsl4
@mohitmthakur please check the invalid DOIs
