### 2026-01-06 ncezid-biome/stylo#8 @telatin
JOSS Review
:link: https://github.com/openjournals/joss-reviews/issues/9695

Hello authors,
happy new year and thanks for sharing *stylo*!
I found the pipeline interesting, and I opened this issue to track my review that will contain both requests (if any) and comments or suggestions that you are free to ignore.

## A. Usage

1. *Suggestion*: The process of updating the lookup table does not look user-friendly. What about adding a column "genome size" in the samplesheet, that can be left empty if the value is present in the lookup table? This would allow executing the pipeline from the repository like "nextflow run ncezid-biome/stylo ..." without tinkering with its files?

2. I see a "test.conf" but this is not linked to the main config: being able to run `nexflow run stylo -profile test,singularity` would be great

## B. Paper

1. Please, remove 'nf-core style' from the title. While it's good to acknowledge the codebase of nf-core in the paper, in the title, this does not really explain anything to the reader and might generate confusion.

2. I think the statement of need needs some improvements, clarifying where the pipeline is positioned for the casual user without needing to know about PulseNet.

4. I think you can put two citations in the same brackets like `[@mbs:/content/journal/mgen/10.1099/mgen.0.001246;@doi:10.1128/jcm.00246-24]` instead of `[@mbs:/content/journal/mgen/10.1099/mgen.0.001246] [@doi:10.1128/jcm.00246-24]`, that would render better in the PDF. Needs to be checked

## C. Other

1. In README.md, substitute `git clone -b nf-core-dev git@github.com:ncezid-biome/stylo.git` with `git clone -b dev git@github.com:ncezid-biome/stylo.git`: branch no longer exists
