### 2026-01-18 ncezid-biome/stylo#15 @mberacochea
Paper missing "State of the Field" - JOSS @mberacochea
The paper doesn't have a "State of the field". How does stylo compare to other pipelines?.
