### 2026-01-18 ncezid-biome/stylo#16 @mberacochea
JOSS Review - @mberacochea
Hi authors,

I've created several issues in the repo after my initial review (https://github.com/openjournals/joss-reviews/issues/9695).

- [ ] https://github.com/ncezid-biome/stylo/issues/9
- [ ] https://github.com/ncezid-biome/stylo/issues/10
- [ ] https://github.com/ncezid-biome/stylo/issues/11
- [ ] https://github.com/ncezid-biome/stylo/issues/12
- [ ] https://github.com/ncezid-biome/stylo/issues/13
- [ ] https://github.com/ncezid-biome/stylo/issues/14
- [ ] https://github.com/ncezid-biome/stylo/issues/15
- [ ] https://github.com/ncezid-biome/stylo/issues/17

Another thing that would be beneficial, probably by using the test data, is to include an example dataset for users to run the pipeline (a small~ish one so users don't have to wait long).
