### 2026-01-18 ncezid-biome/stylo#17 @mberacochea
The pipeline is missing unit tests - JOSS Review @mberacochea
The pipeline is missing unit tests, there is a test profile but it is commented out.
