### 2026-01-18 #9695 @mberacochea
Hi all,

I've completed my initial review, I've compiled the list of tickets I created to keep track [here](https://github.com/ncezid-biome/stylo/issues/16).

Once those are taken care of I will do a second review. @mohitmthakur, I'm happy to discuss any of the issues I created if they don't make sense.

I think the 2 most important ones are:
- https://github.com/ncezid-biome/stylo/issues/15
- https://github.com/ncezid-biome/stylo/issues/17

Cheers
