### 2026-01-20 #9695 @mohitmthakur
Hi @mberacochea ,

Thanks for your initial review! @arzoopatel5 and I discovered an issue last week while running stylo which we are working to resolve asap and will push that update when we've figured it out. Then we will tackle the remaining issues. We will keep you both updated as we resolve them.
