### 2026-01-24 a8d0f3f
fixed medaka bug, hacky version, busco still broken
