+++ modules/nf-core/busco/busco/meta.yml
@@ -40,16 +40,13 @@ input:
   - - config_file:
         type: file
         description: Path to BUSCO config file.
-  - - clean_intermediates:
-        type: boolean
-        description: Flag to remove intermediate files.
 output:
   - batch_summary:
       - meta:
           type: map
           description: |
             Groovy Map containing sample information
-            e.g. [ id:'test' ]
+            e.g. [ id:'test', single_end:false ]
       - "*-busco.batch_summary.txt":
           type: file
           description: Summary of all sequence files analyzed
@@ -59,7 +56,7 @@ output:
           type: map
           description: |
             Groovy Map containing sample information
-            e.g. [ id:'test' ]
+            e.g. [ id:'test', single_end:false ]
       - short_summary.*.txt:
           type: file
           description: Short Busco summary in plain text format
@@ -69,27 +66,17 @@ output:
           type: map
           description: |
             Groovy Map containing sample information
-            e.g. [ id:'test' ]
+            e.g. [ id:'test', single_end:false ]
       - short_summary.*.json:
           type: file
           description: Short Busco summary in JSON format
           pattern: "short_summary.*.json"
-  - log:
-      - meta:
-          type: map
-          description: |
-            Groovy Map containing sample information
-            e.g. [ id:'test' ]
-      - "*-busco/logs/busco.log":
-          type: file
-          description: BUSCO main log
-          pattern: "*-busco/logs/busco.log"
   - full_table:
       - meta:
           type: map
           description: |
             Groovy Map containing sample information
-            e.g. [ id:'test' ]
+            e.g. [ id:'test', single_end:false ]
       - "*-busco/*/run_*/full_table.tsv":
           type: file
           description: Full BUSCO results table
@@ -99,7 +86,7 @@ output:
           type: map
           description: |
             Groovy Map containing sample information
-            e.g. [ id:'test' ]
+            e.g. [ id:'test', single_end:false ]
       - "*-busco/*/run_*/missing_busco_list.tsv":
           type: file
           description: List of missing BUSCOs
@@ -109,7 +96,7 @@ output:
           type: map
           description: |
             Groovy Map containing sample information
-            e.g. [ id:'test' ]
+            e.g. [ id:'test', single_end:false ]
       - "*-busco/*/run_*/single_copy_proteins.faa":
           type: file
           description: Fasta file of single copy proteins (transcriptome mode)
@@ -119,7 +106,7 @@ output:
           type: map
           description: |
             Groovy Map containing sample information
-            e.g. [ id:'test' ]
+            e.g. [ id:'test', single_end:false ]
       - "*-busco/*/run_*/busco_sequences":
           type: directory
           description: BUSCO sequence directory
@@ -129,7 +116,7 @@ output:
           type: map
           description: |
             Groovy Map containing sample information
-            e.g. [ id:'test' ]
+            e.g. [ id:'test', single_end:false ]
       - "*-busco/*/translated_proteins":
           type: directory
           description: Six frame translations of each transcript made by the transcriptome
@@ -140,21 +127,11 @@ output:
           type: map
           description: |
             Groovy Map containing sample information
-            e.g. [ id:'test' ]
+            e.g. [ id:'test', single_end:false ]
       - "*-busco":
           type: directory
           description: BUSCO lineage specific output
           pattern: "*-busco"
-  - downloaded_lineages:
-      - meta:
-          type: map
-          description: |
-            Groovy Map containing sample information
-            e.g. [ id:'test' ]
-      - "busco_downloads/lineages/*":
-          type: directory
-          description: Lineages downloaded by BUSCO when running the analysis, for example bacteria_odb12
-          pattern: "busco_downloads/lineages/*"
   - versions:
       - versions.yml:
           type: file
