+++ modules/nf-core/busco/busco/meta.yml
@@ -1,153 +1,0 @@
-name: busco_busco
-description: Benchmarking Universal Single Copy Orthologs
-keywords:
-  - quality control
-  - genome
-  - transcriptome
-  - proteome
-tools:
-  - busco:
-      description: BUSCO provides measures for quantitative assessment of genome assembly,
-        gene set, and transcriptome completeness based on evolutionarily informed expectations
-        of gene content from near-universal single-copy orthologs selected from OrthoDB.
-      homepage: https://busco.ezlab.org/
-      documentation: https://busco.ezlab.org/busco_userguide.html
-      tool_dev_url: https://gitlab.com/ezlab/busco
-      doi: "10.1007/978-1-4939-9173-0_14"
-      licence: ["MIT"]
-      identifier: biotools:busco
-input:
-  - - meta:
-        type: map
-        description: |
-          Groovy Map containing sample information
-          e.g. [ id:'test', single_end:false ]
-    - fasta:
-        type: file
-        description: Nucleic or amino acid sequence file in FASTA format.
-        pattern: "*.{fasta,fna,fa,fasta.gz,fna.gz,fa.gz}"
-  - - mode:
-        type: string
-        description: The mode to run Busco in. One of genome, proteins, or transcriptome
-        pattern: "{genome,proteins,transcriptome}"
-  - - lineage:
-        type: string
-        description: The BUSCO lineage to use, or "auto", "auto_prok" or "auto_euk"
-          to automatically select lineage
-  - - busco_lineages_path:
-        type: directory
-        description: Path to local BUSCO lineages directory.
-  - - config_file:
-        type: file
-        description: Path to BUSCO config file.
-output:
-  - batch_summary:
-      - meta:
-          type: map
-          description: |
-            Groovy Map containing sample information
-            e.g. [ id:'test', single_end:false ]
-      - "*-busco.batch_summary.txt":
-          type: file
-          description: Summary of all sequence files analyzed
-          pattern: "*-busco.batch_summary.txt"
-  - short_summaries_txt:
-      - meta:
-          type: map
-          description: |
-            Groovy Map containing sample information
-            e.g. [ id:'test', single_end:false ]
-      - short_summary.*.txt:
-          type: file
-          description: Short Busco summary in plain text format
-          pattern: "short_summary.*.txt"
-  - short_summaries_json:
-      - meta:
-          type: map
-          description: |
-            Groovy Map containing sample information
-            e.g. [ id:'test', single_end:false ]
-      - short_summary.*.json:
-          type: file
-          description: Short Busco summary in JSON format
-          pattern: "short_summary.*.json"
-  - full_table:
-      - meta:
-          type: map
-          description: |
-            Groovy Map containing sample information
-            e.g. [ id:'test', single_end:false ]
-      - "*-busco/*/run_*/full_table.tsv":
-          type: file
-          description: Full BUSCO results table
-          pattern: "full_table.tsv"
-  - missing_busco_list:
-      - meta:
-          type: map
-          description: |
-            Groovy Map containing sample information
-            e.g. [ id:'test', single_end:false ]
-      - "*-busco/*/run_*/missing_busco_list.tsv":
-          type: file
-          description: List of missing BUSCOs
-          pattern: "missing_busco_list.tsv"
-  - single_copy_proteins:
-      - meta:
-          type: map
-          description: |
-            Groovy Map containing sample information
-            e.g. [ id:'test', single_end:false ]
-      - "*-busco/*/run_*/single_copy_proteins.faa":
-          type: file
-          description: Fasta file of single copy proteins (transcriptome mode)
-          pattern: "single_copy_proteins.faa"
-  - seq_dir:
-      - meta:
[…]
