### 2026-02-04 62686b5
replace busco with quast in markdown, paper, and tubemap
