### 2026-02-06 #9695 @mohitmthakur
Hello @mberacochea  and @telatin,

@arzoopatel5 and I have been working on stylo for the last few weeks. There was a lapse in appropriations which interrupted our usual workflow late last week and this week but we have resumed. We have been pushing updates to the GitHub as we go along. We fixed the major issue that was affecting the ability to run the pipeline outside of the CDC infrastructure, but that should be resolved now. We still need time on this but wanted to let you know we are still actively working to get as many of your concerns addressed as we can. Thank you.
