+++ README.md
@@ -25,37 +25,32 @@
 > 
 > singularity v3.8.7
 
-First, download this branch to your preferred directory
-```bash
-cd /path/to/dir/
-git clone -b nf-core-dev git@github.com:ncezid-biome/stylo.git
-```
-
-Second, prepare a samplesheet with your input data that looks as follows:
+First, prepare a samplesheet with your input data that looks as follows:
 
 `samplesheet.csv`:
 
 ```csv
-sample,fastq,genus,species
-sample1,/path/to/sample1.fastq.gz,Salmonella,enterica
-sample2,/path/to/sample2.fastq.gz,Campylobacter,coli
-sample3,/path/to/sample3.fastq.gz,Campylobacter,jejuni
-sample4,/path/to/sample4.fastq.gz,Vibrio,-
-sample5,/path/to/sample5.fastq.gz,Salmonella,enterica
+sample,fastq,genus,species,genome_size
+sample1,/path/to/sample1.fastq.gz,Salmonella,enterica,-
+sample2,/path/to/sample2.fastq.gz,Campylobacter,coli,-
+sample3,/path/to/sample3.fastq.gz,Campylobacter,jejuni,-
+sample4,/path/to/sample4.fastq.gz,Vibrio,-,-
+sample5,/path/to/sample5.fastq.gz,Pseudomonas,aeruginosa,6.0m
 ```
 
-Each row represents a fastq file (single-end) with the known genus and species.
+Each row represents a fastq file (single-end) with the known genus, species, and genome size.
 
 > [!NOTE]
-> you can use `-` where the species is unknown
+> You can use `-` where the species is unknown. If the organism is not in the [lookup table](conf/lookup_table.tsv), you must include a genome size. for more details see [full samplesheet description](docs/usage.md#full-samplesheet)
 
-Third, look at the [lookup table](conf/lookup_table.tsv) to make sure that each genus listed in your samplesheet is present. If you'd like to add a row or edit the lookup table see [Advanced Usage](#advanced-usage)
+If you'd like to edit the lookup table to add your organism as a default see [Advanced Usage](docs/usage.md#advanced-usage)
 
 
 Now, you can run the pipeline using:
 
 ```bash
-nextflow run /path/to/stylo/main.nf \
+nextflow run ncezid-biome/stylo \
+   -r v1.3.0 \
    -profile singularity
    --input samplesheet.csv \
    --outdir <OUTDIR>
@@ -68,73 +63,7 @@ If you're on CDC servers use `-profile rosalind`
 > Please provide pipeline parameters via the CLI or Nextflow `-params-file` option. Custom config files including those provided by the `-c` Nextflow option can be used to provide any configuration _**except for parameters**_;
 > see [docs](https://nf-co.re/usage/configuration#custom-configuration-files).
 
-For more details about generic usage see the [Usage Page](docs/usage.md)
-
-## Advanced Usage
-
-### Editing the Lookup Table
-If a genus is missing, then you'll need to add a row to the lookup table prior to running the pipeline. In order to add a row to the lookup table you'll need the following information:
-
-1. genus (required)
-2. species (optional, use `-` if you want the lookup table to accept all species within that genus)
[…]
