+++ docs/usage.md
@@ -12,11 +12,11 @@ You will need to create a samplesheet with information about the samples you wou
 
 ### Full samplesheet
 
-The samplesheet can have as many columns as you desire, however, there is a strict requirement for the first 4 columns to match those defined in the table below.
+There is a strict requirement for the first 4 columns of the samplesheet to match those defined in the table below. The 5th column is optional.
 
-A final samplesheet file may look something like the one below.
+A few example samplesheets are provided below and included in [assets](../assets/).
 
-```csv title="samplesheet.csv"
+```csv title="samplesheet_basic.csv"
 sample,fastq,genus,species
 sample1,/path/to/sample1.fastq.gz,Salmonella,enterica
 sample2,/path/to/sample2.fastq.gz,Campylobacter,coli
@@ -25,12 +25,22 @@ sample4,/path/to/sample4.fastq.gz,Vibrio,-
 sample5,/path/to/sample5.fastq.gz,Salmonella,enterica
 ```
 
-| Column    | Description                                                                                                                                                                            |
-| --------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| `sample`  | Custom sample name. This entry must be unique. |
-| `fastq` | Full path to FastQ file for ONT longreads. File has to be gzipped and have the extension ".fastq.gz" or ".fq.gz".                                                             |
-| `genus` | genus of the sample. This must be provided for the pipeline to run, otherwise the row will be skipped. |
-| `species` | species of the sample. If you don't know the species or would like to skip this part use `-` as seen in the example samplesheet. Note that this might affect some assemblies such as Vibrio where different species within the genus have different genome sizes |
+```csv title="samplesheet_extra_genomes.csv"
+sample,fastq,genus,species,genome_size
+sample1,/path/to/sample1.fastq.gz,Salmonella,enterica,4.8m
+sample2,/path/to/sample2.fastq.gz,Campylobacter,coli,-
+sample3,/path/to/sample3.fastq.gz,Pseudomonas,aeruginosa,6.0m
+sample3,/path/to/sample3.fastq.gz,Bacteroides,fragilis,5.2m
+```
+
+
+| Column    | Format | Description                                                                                                                                                                            |
+| --------- | ------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| `sample`  | string | Custom sample name. This entry must be unique. |
+| `fastq` | path | Full path to FastQ file for ONT longreads. File has to be gzipped and have the extension ".fastq.gz" or ".fq.gz".                                                             |
+| `genus` | string | genus of the sample. This must be provided for the pipeline to run, otherwise the row will be skipped. |
+| `species` | string OR `-` | species of the sample. If you don't know the species or would like to skip this part use `-` as seen in the example samplesheet. Note that this might affect some assemblies such as Vibrio where different species within the genus have different genome sizes |
[…]
