### 2026-02-20 8b5d586
cleaned up config files, fixed min overlap, updated README, added CDC.md, removed gzip restriction, added option for lookup table. fix #9, fix #10, fix #13, fix #14
