### 2026-02-20 dc50345
update wording in READMEs
