+++ .github/workflows/plasmid.yml
@@ -4,6 +4,7 @@ jobs:
   example:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - uses: nf-core/setup-nextflow@v1
+      - uses: eWaterCycle/setup-singularity@v7
       - run: nextflow run main.nf -profile test,singularity --outdir output/
