+++ .github/workflows/plasmid.yml
@@ -1,10 +1,16 @@
 name: plasmid workflow
 on: push
 jobs:
-  example:
+  plasmid:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v4 
       - uses: nf-core/setup-nextflow@v1
-      - uses: eWaterCycle/setup-singularity@v7
-      - run: nextflow run main.nf -profile test,singularity --outdir output/
+      - uses: eWaterCycle/setup-apptainer@4bb22c52d4f63406c49e94c804632975787312b3
+        if: matrix.profile == 'singularity'
+        run: |
+          mkdir -p $NXF_SINGULARITY_CACHEDIR
+          mkdir -p $NXF_SINGULARITY_LIBRARYDIR
+      - run: nextflow 
+          run main.nf -profile test,singularity --outdir output/
+ 
