### 2026-02-23 c5b9993
added test to README
