+++ .github/workflows/plasmid.yml
@@ -1,12 +1,0 @@
-name: plasmid workflow
-on: push
-jobs:
-  plasmid:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4 
-      - uses: nf-core/setup-nextflow@v1
-      - uses: eWaterCycle/setup-apptainer@4bb22c52d4f63406c49e94c804632975787312b3
-        if: matrix.profile == 'singularity'
-      - run: nextflow run main.nf -profile test,singularity --outdir output/
- 
