### 2026-02-26 49d4fa6
implemented MultiQC for nanoq and quast and added references to report. fix #11
