+++ assets/multiqc_config.yml
@@ -8,6 +8,22 @@ report_section_order:
     order: -1001
   "ncezid-biome-stylo-summary":
     order: -1002
+  nanoq:
+    order: 2000
+  quast:
+    order: 1000
+
+quast_config:
+  contig_length_multiplier: 0.000001
+  contig_length_suffix: "Mbp"
+  total_length_multiplier: 0.000001
+  total_length_suffix: "Mbp"
+  total_number_contigs_multiplier: 1
+  total_number_contigs_suffix: ""
+
+sample_names_replace_regex: True
+sample_names_replace:
+  "(\\S)_(filtered)": "\\1"
 
 export_plots: true
 
