+++ paper/paper.md
@@ -1,5 +1,5 @@
 --- 
-title: "stylo: a lightweight nf-core style nanopore assembly pipeline optimized for enteric bacteria"
+title: "stylo: a lightweight nanopore assembly pipeline optimized for enteric bacteria"
 tags: 
   - nextflow 
   - bioinformatics 
@@ -54,9 +54,9 @@ Oxford Nanopore Technologies (ONT) sequencing is a promising technology with man
 
 # Statement of Need
 
-There is a continuous need for foodborne outbreak detection in public health. To determine the scope or severity of a foodborne outbreak, short-read whole genome sequencing has often been used to generate isolate assemblies of enteric bacteria often which supports rapid and accurate outbreak detection [@Ribot2016-ps] [@Timme2017-gl]. However, as nanopore long-read sequencing becomes more cost-effective and accurate, the need increases for streamlined assembly pipelines to support high-throughput surveillance processing of ONT sequenced isolates [@mbs:/content/journal/mgen/10.1099/mgen.0.001246] [@doi:10.1128/jcm.00246-24]. With the increased adoption of modern high-performance computing and cloud servers, pipelines built to leverage containerization and custom configurations allow for easy deployment on those servers. To address these needs, we have created stylo, a lightweight nf-core style nanopore assembly pipeline optimized for enteric bacteria [@Di-Tommaso2017] [@Ewels2020].  
+There is a continuous need for foodborne outbreak detection in public health. To determine the scope or severity of a foodborne outbreak, short-read whole genome sequencing has often been used to generate isolate assemblies of enteric bacteria often which supports rapid and accurate outbreak detection [@Ribot2016-ps;@Timme2017-gl]. However, as nanopore long-read sequencing becomes more cost-effective and accurate, the need increases for streamlined assembly pipelines to support high-throughput surveillance processing of ONT sequenced isolates [@mbs:/content/journal/mgen/10.1099/mgen.0.001246;@doi:10.1128/jcm.00246-24]. With the increased adoption of modern high-performance computing and cloud servers, pipelines built to leverage containerization and custom configurations allow for easy deployment on those servers. To address these needs, we have created stylo, a lightweight nf-core style nanopore assembly pipeline optimized for enteric bacteria [@Di-Tommaso2017;@Ewels2020].  
 
-Stylo is developed for PulseNet, a molecular surveillance network for foodborne infections in the United States [@doi:10.1089/fpd.2006.3.9]. PulseNet facilitates the rapid detection of illness clusters and reduces the likelihood of outbreaks becoming large and widespread [@doi:10.1089/fpd.2019.2637]. Stylo utilizes a lookup table of PulseNet organism genome sizes, allowing users to run large and diverse datasets of enteric bacteria. There exists generalized nanopore assembly workflows such as Donut Falls [@DonutFalls2025], whereas stylo is a streamlined workflow specialized for PulseNet to facilitate downstream genotyping.  
+Since stylo is lightweight, it can be run in automated disease surveillance settings, such as PulseNet, which facilitates the rapid detection of illness clusters and reduces the likelihood of outbreaks becoming large and widespread [@doi:10.1089/fpd.2006.3.9;@doi:10.1089/fpd.2019.2637]. Stylo was designed and tested around enteric pathogens and default genome sizes were obtained through PulseNet Standard Operating Procedure [@PulseNet-sop2025], however, it does not have any strict limitation to enteric bacteria. This allows users to customize the lookup table to include any bacteria that they work with. There exist extensive nanopore assembly workflows such as Donut Falls [@DonutFalls2025], which includes hybrid assembly options and multiple rounds of polishing. However, stylo is a streamlined nanopore workflow, intended for high throughput research groups, surveillance networks, or practical assemblies for general users.
 
 # Workflow Overview
 
