+++ paper/paper.bib
@@ -195,3 +195,62 @@ eprint = {https://doi.org/10.1089/fpd.2019.2637}
   journal = {GitHub repository},
   url = {https://github.com/UPHL-BioNGS/Donut_Falls}
 }
+
+@article{doi:10.1093/bioinformatics/btw354,
+ author = {Ewels, Philip and Magnusson, Måns and Lundin, Sverker and Käller, Max},
+ title = {MultiQC: summarize analysis results for multiple tools and samples in a single report},
+ journal = {Bioinformatics},
+ volume = {32},
+ number = {19},
+ pages = {3047},
+ year = {2016},
+ doi = {10.1093/bioinformatics/btw354},
+ URL = { + http://dx.doi.org/10.1093/bioinformatics/btw354},
+ eprint = {/oup/backfile/Content_public/Journal/bioinformatics/32/19/10.1093_bioinformatics_btw354/3/btw354.pdf}
+}
+
+@article{10.1093/bioinformatics/btx192,
+    author = {da Veiga Leprevost, Felipe and Grüning, Björn A and Alves Aflitos, Saulo and Röst, Hannes L and Uszkoreit, Julian and Barsnes, Harald and Vaudel, Marc and Moreno, Pablo and Gatto, Laurent and Weber, Jonas and Bai, Mingze and Jimenez, Rafael C and Sachsenberg, Timo and Pfeuffer, Julianus and Vera Alvarez, Roberto and Griss, Johannes and Nesvizhskii, Alexey I and Perez-Riverol, Yasset},
+    title = {BioContainers: an open-source and community-driven framework for software standardization},
+    journal = {Bioinformatics},
+    volume = {33},
+    number = {16},
+    pages = {2580-2582},
+    year = {2017},
+    month = {03},
+    issn = {1367-4803},
+    doi = {10.1093/bioinformatics/btx192},
+    url = {https://doi.org/10.1093/bioinformatics/btx192},
+    eprint = {https://academic.oup.com/bioinformatics/article-pdf/33/16/2580/49041124/bioinformatics_33_16_2580.pdf},
+}
+
+@article{10.1093/nar/gkae410,
+    author = {The Galaxy Community },
+    title = {The Galaxy platform for accessible, reproducible, and collaborative data analyses: 2024 update},
+    journal = {Nucleic Acids Research},
+    volume = {52},
+    number = {W1},
+    pages = {W83-W94},
+    year = {2024},
+    month = {05},
+    issn = {0305-1048},
+    doi = {10.1093/nar/gkae410},
+    url = {https://doi.org/10.1093/nar/gkae410},
+    eprint = {https://academic.oup.com/nar/article-pdf/52/W1/W83/58436305/gkae410.pdf},
+}
+
+@misc{seqera-wave,
+  title = {Wave},
+  year = {2022},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/seqeralabs/wave}
+}
+
+@misc{PulseNet-sop2025,
+  title = {PulseNet International SOPs},
+  year = {2025},
+  publisher = {APHL},
+  journal = {APHL repository},
+  url = {https://www.aphl.org/programs/global_health/Pages/PulseNet-International-SOPs.aspx}
+}
