### 2026-02-27 ff772c9
added multiqc to tubemap and paper
