### 2026-02-27 ncezid-biome/stylo#8 @mohitmthakur
Hello @telatin ,

We have addressed your concerns and the current version of stylo has been pushed. We have one citation which we need to verify but otherwise the main branch is now up to date with the dev. Please let us know if you have any issues with running the pipeline using the test case we've provided (see ReadMe) and also if there is anything else you encounter in your second pass. Thank you!

```
nextflow run ncezid-biome/stylo \
   -r v1.3.2 \
   -profile test,singularity \
   --outdir stylo_test/
```
