### 2026-03-06 ncezid-biome/stylo#15 @arzoopatel5
Sorry, you're right, we noticed that when editing the paper. Since our last comment on this thread, we've added some details about donut falls and how it compares to stylo. We believe this is sufficient, however we can expand on it and/or separate it into its own section if you'd like.
