+++ README.md
@@ -8,7 +8,7 @@
 2. Downsamples reads to specific coverage ([rasusa](https://github.com/mbhall88/rasusa))
 3. Assembles reads ([Flye](https://github.com/mikolmogorov/Flye))
 4. Reorients assembly ([Dnaapler](https://github.com/gbouras13/dnaapler))
-5. Error correction ([Medaka](https://github.com/nanoporetech/medaka))
+5. Error correction ([Dorado polish](https://github.com/nanoporetech/dorado))
 6. QCs assembly ([QUAST](https://github.com/ablab/quast))
 
 ## Usage
