+++ assets/multiqc_config.yml
@@ -24,6 +24,7 @@ quast_config:
 sample_names_replace_regex: True
 sample_names_replace:
   "(\\S)_(filtered)": "\\1"
+  "(\\S)_(polished)": "\\1"
 
 export_plots: true
 
