### 2026-03-17 #9695 @mberacochea
Hi @mohitmthakur, I bumped into some issues while trying to test the pipeline locally - I've tried to fix some of the errors here - https://github.com/ncezid-biome/stylo/pull/22. There is still an issue with Medaka which I wasn't able to fix.
