### 2026-03-17 a817d2c
fixed logging, updated docs
