### 2026-03-17 #9695 @arzoopatel5
Hi @mberacochea @telatin, just for your awareness, we've swapped medaka to dorado polish for some efficiency improvements and made edits to the documentation and paper to reflect the change.
