### 2026-03-23 #9695 @mberacochea
I've sent you folks a PR that includes a whole-pipeline nf-test. I will try to finish my review this week. @mohitmthakur @arzoopatel5
