### 2026-03-28 #9695 @mberacochea
### Software review

I'm done with my review of the code, and it looks good to me (thanks for addressing my comments).

Recommendations — non blockers

Enabling CI using GitHub Actions would be good, but Dorado makes it a difficult because the container is heavy. A stop-gap solution would be to add small section in the README about the pipeline's end-to-end testing (how to run this).

### Software paper review

I think the paper is well written and covers the following JOSS-required sections effectively:
- Statement of need
- Software design
- Research impact statement

The section that is missing, or not covered in enough detail, is the "State of the field". I'll defer to @crsl4 about this
