### 2026-04-13 #9695 @arzoopatel5
Hi @crsl4, could you let us know what your thoughts are on the "State of the field" section?
