### 2026-05-12 #9695 @jchen232
Hi @arfon — apologies for the tag, but we wanted to flag this review for your attention having had some difficulty reaching the editor.

A quick summary of where things stand:

- **Both reviewers have completed their checklists** — @telatin signed off on March 6 and @mberacochea on March 28
- The one open editorial question — coverage of the "State of the field" — was raised by @mberacochea, who deferred it to crsl4. We flagged it here on April 28 and again two weeks ago, but haven't heard back yet. This section wasn't required at the time of submission, but our team added some information under statement of need.

We don't want to be a burden, and we're genuinely grateful to @mberacochea and @telatin for their thorough and constructive reviews — stylo is a much stronger package as a result. We just want to make sure their efforts don't go unrecognized.

Could you advise on the best next steps? We're happy to make any further changes needed on our end. Thank you so much for your time.
